feat: Pass file path to getContext
The `getContext` function now accepts a `path` argument, which is stored in the `Context` object. This path is then used in `updateRulesList` to correctly determine the plugin root when generating error messages related to missing rule list markers. Previously, this was done using `pathPlugin`, which was removed.

The `pathPlugin` argument has been removed from `generateRuleHeaderLines` and `updateRulesList`.

(This commit message was automatically generated by an LLM.)